+++ examples/06-transformations/src/TransformationsRenderer.hpp
@@ -13,7 +13,7 @@ class TransformationsRenderer : public Renderer
 public:
     virtual bool onInit() override;
     virtual void onDestroy() override;
-    virtual void onUpdate(double time) override;
+    virtual void onUpdate(float time) override;
     virtual void onDraw() override;
 private:
     ShaderProgram* m_shader = nullptr;
